backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.3.7
anthropic>=0.7.7
google-generativeai>=0.3.1
pydantic-ai>=0.0.5
httpx>=0.25.2
logfire>=0.8.0
devtools>=0.12.2
streamlit>=1.29.0
